feat: extract hdwalletNativeVaultsList queryKey / queryFn
