+++ src/context/WalletProvider/NativeWallet/components/NativeSuccess.tsx
@@ -1,5 +1,6 @@
 import { Box, ModalBody, ModalHeader } from '@chakra-ui/react'
 import { useQueryClient } from '@tanstack/react-query'
+import { reactQueries } from 'react-queries'
 import { Text } from 'components/Text'
 
 import { useNativeSuccess } from '../hooks/useNativeSuccess'
@@ -9,7 +10,9 @@ export const NativeSuccess = ({ location }: NativeSetupProps) => {
   const queryClient = useQueryClient()
   const { isSuccessful } = useNativeSuccess({ vault: location.state.vault })
 
-  queryClient.invalidateQueries({ queryKey: ['hdwalletNativeVaultsList'] })
+  queryClient.invalidateQueries({
+    queryKey: reactQueries.common.hdwalletNativeVaultsList().queryKey,
+  })
 
   return (
     <>
